Update .gitignore and remove .idea directory
The .idea directory should not be stored in Git,
as it's user-specific and part of the project itself.

It was there due to incorrect .gitignore file,
so in this commit that file was updated.

Maven-specific entries were included alongside
Gradle-specific ones, as I want to migrate to Maven
in the near future.