Use plain Java instead of Spring Boot by chess-server
The communication model will be using sockets for communication,
because it is bidirectional by design and there is no need to
simulate this behaviour by sending server-to-client messages
in HTTP responses etc.

That means that I have to manage my socket connection directly
and thus there is no use in using Spring Boot in that project.