God-commit: Refactor structure of whole server
I've decided to restructure the whole server.

The resources are managed by as small number of classes as possible
(preferably 1), the error handling and shutdown process was simplified
to follow the tree-like structure of the threads, the communication
model was simplified, for example new piece type needs to be sent
together with the promotion move to avoid additional message exchange.

MatchRegister was implemented. It keeps track of open connections,
matches players into games, and clean up closed or otherwise dead
connections. It also keeps track of active games, and terminates
them at shutdown.