Decouple client from model's internals - Part 3
PieceType was moved to package ....model.properties,
so in the future it can be exposed from the client module,
while the domain representation of the pieces can remain
encapsulated inside the model module.

Additionally, the dependency on the Game object was moved
from BoardScreenController to WelcomeScreenView,
so it can pick whichever one is appropriate
to users decision in the menu screen.