Store primary stage as a global variable
Primary stage is central object in JavaFX application.
I decided to make it a read-only global variable to avoid
passing it in circles as parameter to most controllers.

I feel the decision is justified for a couple of reasons:

1. To ensure the most smooth user experience it's better
to reuse one existing window instead of always closing
the previous one and opening the next.
Thus, most of the controllers need reference to it anyway.
2. Concurrent modification is not a problem, because GUI
can be only modified from JavaFX Application Thread.
3. The variable is read-only, so we cannot lose/change
the Stage it's referring to but only interact with main window.